Reversion history in sidebar instad of popal
