+++ include/Page/Edit.php
@@ -99,7 +99,6 @@ class Edit extends \gp\Page{
 			$this->cmds['UseRevision']				= 'DefaultDisplay';
 			$this->cmds['ViewHistory']				= '';
 			$this->cmds['ViewCurrent']				= '';
-			$this->cmds['DeleteRevision']			= 'ViewHistory';
 			$this->cmds['PublishDraft']				= 'DefaultDisplay';
 
 			/* inline editing */
@@ -191,13 +190,12 @@ class Edit extends \gp\Page{
 		}
 
 		$admin_links[]	= \gp\tool::Link(
-			$this->title,
+			'/Admin/Revisions/'.$this->gp_index,
 			'<i class="fa fa-history"></i> ' . $langmessage['Revision History'],
 			'cmd=ViewHistory',
 			array(
 				'title'		=> $langmessage['Revision History'],
 				'class'		=> 'admin-link admin-link-revision-history',
-				'data-cmd'	=> 'gpabox',
 			)
 		);
 
@@ -381,12 +379,11 @@ class Edit extends \gp\Page{
 		}
 
 		$admin_links[] = \gp\tool::Link(
-			$this->title,
+			'/Admin/Revisions/'.$this->gp_index,
 			'<i class="fa fa-history"></i> ' . $langmessage['Revision History'],
 			'cmd=ViewHistory',
 			array(
 				'title'		=> $langmessage['Revision History'],
-				'data-cmd'	=> 'gpabox'
 			)
 		);
 
@@ -1431,53 +1428,6 @@ class Edit extends \gp\Page{
 
 
 	/**
-	 * Display the revision history of the current file
-	 *
-	 */
-	public function ViewHistory(){
-		global $langmessage, $config;
-
-		$files		= $this->BackupFiles();
-		$rows		= array();
-
-		//working draft
-		if( $this->draft_exists ){
-			$draft_file = \gp\tool\Files::FilePath($this->draft_file);
-			$size = filesize($draft_file);
-			$time = $this->file_stats['modified'];
-			$rows[$time] = $this->HistoryRow($time, $size, $this->file_stats['username'], 'draft');
-		}
-
-		foreach($files as $time => $file){
-			$info = $this->BackupInfo($file);
-			$rows[$time] = $this->HistoryRow($info['time'], $info['size'], $info['username']);
-		}
-
-		// current page
-		// this will overwrite one of the history entries if there is a draft
-		$page_file = \gp\tool\Files::FilePath($this->file);
-		$rows[$this->fileModTime] = $this->HistoryRow($this->fileModTime, filesize($page_file), $this->file_stats['username'], 'current');
-
-		echo '<h2>' . $langmessage['Revision History'] . '</h2>';
-		echo '<table class="bordered full_width striped"><tr>';
-		echo '<th>' . $langmessage['Modified'] . '</th>';
-		echo '<th>' . $langmessage['File Size'] . '</th>';
-		echo '<th>' . $langmessage['username'] . '</th>';
-		echo '<th>&nbsp;</th>';
-		echo '</tr><tbody>';
-
-		krsort($rows);
-		echo implode('', $rows);
-
-		echo '</tbody>';
-		echo '</table>';
-
-		echo '<p>' . $langmessage['history_limit'] . ': ' . $config['history_limit'] . '</p>';
-	}
-
-
-
-	/**
 	 * Get info about a backup from the filename
 	 *
 	 */
@@ -1505,98 +1455,22 @@ class Edit extends \gp\Page{
 	}
 
 
-
-	/**
-	 * Return content for history row
-	 *
-	 */
-	protected function HistoryRow($time, $size, $username, $which='history'){
-		global $langmessage;
-
-		ob_start();
-		$date = \gp\tool::date($langmessage['strftime_datetime'], $time);
-		echo '<tr><td title="' . htmlspecialchars($date) . '">';
-		switch($which){
-			case 'current':
-				echo '<b>' . $langmessage['Current Page'] . '</b><br/>';
-				break;
-
-			case 'draft':
-				echo '<b>' . $langmessage['Working Draft'] . '</b><br/>';
-				break;
-		}
-
-		$elapsed = \gp\admin\Tools::Elapsed(time() - $time);
-		echo sprintf($langmessage['_ago'], $elapsed);
-		echo '</td><td>';
-		if( $size && is_numeric($size) ){
-			echo \gp\admin\Tools::FormatBytes($size);
-		}
-		echo '</td><td>';
-		if( !empty($username) ){
-			echo $username;
-		}
-		echo '</td><td>';
-
-		switch($which){
-			case 'current':
-				echo \gp\tool::Link(
-					$this->title,
-					$langmessage['View'],
-					'cmd=ViewCurrent',
-					array('data-cmd' => 'cnreq')
-				);
-				break;
-
-			case 'draft':
-				echo \gp\tool::Link($this->title, $langmessage['View']);
-				echo ' &nbsp; ' . \gp\tool::Link(
-					$this->title,
-					$langmessage['Publish Draft'],
-					'cmd=PublishDraft',
-					array('data-cmd' => 'cnreq')
-				);
-				break;
-
-			case 'history':
-				echo \gp\tool::Link(
-					$this->title,
-					$langmessage['View'],
-					'cmd=ViewRevision&time=' . $time,
-					array('data-cmd' => 'cnreq')
-				);
-				echo ' &nbsp; ';
-				echo \gp\tool::Link(
-					$this->title,
-					$langmessage['delete'],
-					'cmd=DeleteRevision&time=' . $time,
-					array(
-						'data-cmd'	=> 'gpabox',
-						'class'		=> 'gpconfirm'
-					)
-				);
-				break;
-		}
-
-		echo '</td></tr>';
-		return ob_get_clean();
-	}
-
-
-
 	/**
 	 * Display the contents of a past revision
 	 *
 	 */
 	protected function ViewRevision(){
-		global $langmessage;
 
-		$time			=& $_REQUEST['time'];
-		$file_sections	= $this->GetRevision($time);
+		\gp\admin\Tools::$show_toolbar	= false;
+		$time							=& $_REQUEST['time'];
+		$file_sections					= $this->GetRevision($time);
+
 
 		if( $file_sections === false ){
-			return false;
+			$this->DefaultDisplay();
+			return;
 		}
+
 
 		$this->revision			= $time;
 		echo \gp\tool\Output\Sections::Render($file_sections, $this->title, \gp\tool\Files::$last_stats);
@@ -1609,7 +1483,6 @@ class Edit extends \gp\Page{
 	 *
 	 */
 	protected function UseRevision(){
-		global $langmessage;
 
 		$time			=& $_REQUEST['time'];
 		$file_sections	= $this->GetRevision($time);
@@ -1665,6 +1538,8 @@ class Edit extends \gp\Page{
 	 */
 	public function ViewCurrent(){
 
+		\gp\admin\Tools::$show_toolbar		= false;
+
 		if( !$this->draft_exists ){
 			$this->DefaultDisplay();
 			return;
@@ -1676,23 +1551,6 @@ class Edit extends \gp\Page{
 	}
 
 
-
-	/**
-	 * Delete a revision backup
-	 *
-	 */
-	public function DeleteRevision(){
-		global $langmessage;
-
-		$full_path	= $this->BackupFile($_REQUEST['time']);
-		if( is_null($full_path) ){
-			return false;
-		}
-		unlink($full_path);
-	}
-
-
-
 	/**
 	 * Return a list of the available backup for the current file
 	 *
@@ -1700,6 +1558,7 @@ class Edit extends \gp\Page{
 	public function BackupFiles(){
 		global $dataDir;
 		$dir = $dataDir . '/data/_backup/pages/' . $this->gp_index;
+
 		if( !file_exists($dir) ){
 			return array();
 		}
@@ -1845,10 +1704,9 @@ class Edit extends \gp\Page{
 						array('data-cmd' => 'gpajax')
 					);
 				echo \gp\tool::Link(
-						$this->title,
+						'/Admin/Revisions/'.$this->gp_index,
 						$langmessage['Revision History'],
-						'cmd=ViewHistory',
-						array( 'data-cmd' => 'gpabox')
+						'cmd=ViewHistory'
 					);
 				echo '<span class="gp_separator"></span>';
 				echo \gp\tool::Link('Admin/Menu', $langmessage['file_manager']);
